+++ navigation/AuthStack.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import LoginScreen from "../screens/LoginScreen";
+import SignUpScreen from "../screens/SignUpScreen";
+import OTPScreen from "../screens/OTPScreen";
+
+const Stack = createNativeStackNavigator(); 
+
+export default function AuthStack() {
+    return (
+        <Stack.Navigator> 
+          <Stack.Screen name="Login" component={LoginScreen} options={{ headerShown: false }} />
+          <Stack.Screen name="SignUp" component={SignUpScreen} options={{ headerShown: false }} />
+          <Stack.Screen name="OTP" component={OTPScreen} options={{ headerShown: false}} />
+        </Stack.Navigator>
+    );
+
+}
