+++ tailwind.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   content: [
     "./App.{js,jsx,ts,tsx}",
-    "./src/**/*.{js,jsx,ts,tsx}",
+    "./screens/*.{js,jsx,ts,tsx}",
     "./index.{js,jsx,ts,tsx}",
     "./components/**/*.{js,jsx,ts,tsx}",
     "./LoginScreen.{js,jsx,ts,tsx}",
